pyproject.toml
[project]
name = "umamusume-auto-train"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "click==8.2.1",
    "colorama==0.4.6",
    "dxcam==0.0.5; sys_platform == 'linux'",
    "dxcam-cpp==0.2.4; sys_platform == 'win32'",
    "easyocr==1.7.2",
    "exceptiongroup==1.3.0",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "fsspec==2025.7.0",
    "h11==0.16.0",
    "idna==3.10",
    "imageio==2.37.0",
    "jinja2==3.1.6",
    "keyboard==0.13.5",
    "lazy-loader==0.4",
    "levenshtein==0.27.1",
    "markupsafe==3.0.2",
    "mouseinfo==0.1.3",
    "mpmath==1.3.0",
    "mss==10.0.0",
    "networkx==3.4.2",
    "ninja==1.11.1.4",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "opencv-python-headless==4.12.0.88",
    "packaging==25.0",
    "pillow>=12.0.0",
    "pyautogui==0.9.54",
    "pyclipper==1.3.0.post6",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygetwindow==0.0.9",
    "pymsgbox==1.0.9",
    "pyperclip==1.9.0",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "python-bidi==0.6.6",
    "pytweening==1.2.0",
    "pyyaml==6.0.2",
    "rapidfuzz==3.13.0",
    "scikit-image==0.25.2",
    "scipy==1.15.3",
    "shapely==2.1.1",
    "sniffio==1.3.1",
    "starlette==0.47.2",
    "sympy==1.14.0",
    "tifffile==2025.5.10",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision==0.24.0+cu130; sys_platform == 'win32'",
    "torchvision==0.24.0+cpu; sys_platform == 'linux'",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "uvicorn==0.35.0",
    "msgspec>=0.19.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true


[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform == 'linux'" },
  { index = "pytorch-cu130", marker = "sys_platform == 'win32'" },
]
